Rename local thread live writer module
